vanityapp/backend/requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
solana==0.30.2
solders==0.18.1
base58==2.1.1
aiohttp==3.9.1

fastapi
uvicorn[standard]
python-dotenv
requests
pydantic